refactor(renderer): modularize spatial_renderer god file
Split the 2619-line spatial_renderer.rs into a directory module without
any change to production logic (same code, relocated):

- spatial_renderer/mod.rs        SpatialRenderer + render_frame + construction (1814 l)
- spatial_renderer/components.rs  support types & free fns: RenderedFrame,
                                  SpatialChannelEvent (pub, re-exported),
                                  ChannelState, BandRenderer, split_bands,
                                  evaluation_build_config (pub(super))
- spatial_renderer/tests.rs       the #[cfg(test)] module

LiveSnapshot stays in mod.rs (its 18 fields are render_frame's intrinsic
working set; moving it would force pub(super) on every field for no gain).
Public paths preserved via `pub use components::{RenderedFrame,
SpatialChannelEvent}`.

Hot path untouched: the render_frame body is byte-for-byte the same and
relocating items within a crate does not change codegen. Verified with the
render_frame Criterion bench — deltas (-2.6%..+4.5%) are run-to-run noise:
a zero-change re-run swung crossover/interp by -11.8% on its own. 95/95
renderer lib tests pass; build + fmt clean.
