+++ omniphony-renderer/renderer/src/spatial_renderer/mod.rs
@@ -71,13 +71,11 @@ use crate::live_params::{
     PolarEvaluationParams, RampMode, RenderTopology, RendererControl,
 };
 use crate::ramp_strategy::{
-    ChannelRampState, PositionRampStrategy, RampContext, RampProgress, RampRenderParams,
-    RampStrategy, RampTarget,
+    PositionRampStrategy, RampContext, RampProgress, RampRenderParams, RampStrategy, RampTarget,
 };
 use crate::render_backend::RenderRequest;
 use crate::render_backend::{
-    CartesianEvaluationConfig, EffectiveEvaluationMode, EvaluationBuildConfig, MultiBandTable,
-    PolarEvaluationConfig, PreparedRenderEngine, VbapBackend, build_prepared_render_engine,
+    EffectiveEvaluationMode, MultiBandTable, VbapBackend, build_prepared_render_engine,
 };
 use crate::spatial_vbap::{DistanceModel, Gains};
 use crate::spatial_vbap::{VbapPanner, VbapTableMode};
@@ -85,278 +83,9 @@ use crate::speaker_layout::SpeakerLayout;
 use anyhow::Result;
 use std::sync::Arc;
 
-/// Output of a single rendered frame.
-pub struct RenderedFrame {
-    /// Interleaved speaker audio: `[sample0_spk0, sample0_spk1, ..., sample1_spk0, ...]`.
-    pub samples: Vec<f32>,
-    /// VBAP gains at the final sample for each rendered object channel.
-    /// `(channel_idx, gains)` — `gains[speaker_idx]` is the gain applied to that speaker.
-    /// Ordered by `channel_idx`. Empty if no objects were spatialized this frame.
-    pub object_gains: Vec<(usize, Gains)>,
-    /// Per-band VBAP gains for crossover objects.
-    /// `(channel_idx, [band0_gains, band1_gains, ...])` — each `Gains` is full-size
-    /// (`num_speakers`), indexed by global speaker index.
-    /// Empty for non-crossover objects or when no crossover is active.
-    pub object_band_gains: Vec<(usize, Vec<Gains>)>,
-    /// Time spent in the crossover filter-bank stage during `render_frame`.
-    ///
-    /// This is a subset of the total render time.
-    pub crossover_time_ms: f32,
-}
-
-fn evaluation_build_config(
-    request_template: RenderRequest,
-    position_interpolation: bool,
-    table_mode: VbapTableMode,
-    azimuth_resolution: i32,
-    elevation_resolution: i32,
-    distance_step: f32,
-    distance_max: f32,
-    allow_negative_z: bool,
-) -> EvaluationBuildConfig {
-    let cartesian = match table_mode {
-        VbapTableMode::Cartesian {
-            x_size,
-            y_size,
-            z_size,
-            z_neg_size,
-        } => CartesianEvaluationConfig {
-            x_size,
-            y_size,
-            z_size,
-            z_neg_size,
-        },
-        VbapTableMode::Polar => CartesianEvaluationConfig {
-            x_size: 2,
-            y_size: 2,
-            z_size: 2,
-            z_neg_size: usize::from(allow_negative_z),
-        },
-    };
-    let azimuth_values = (360 / azimuth_resolution.max(1)).max(2) as usize;
-    let elevation_span = if allow_negative_z { 180 } else { 90 };
-    let elevation_values = (elevation_span / elevation_resolution.max(1)).max(2) as usize;
-    let distance_values =
-        ((distance_max.max(0.01) / distance_step.max(0.01)).round() as usize).max(1) + 1;
-    EvaluationBuildConfig {
-        request_template,
-        position_interpolation,
-        cartesian,
-        polar: PolarEvaluationConfig {
-            azimuth_values,
-            elevation_values,
-            distance_values,
-            distance_max,
-            allow_negative_z,
-        },
-        // Initial build uses default metrics; a configured non-default metric is
-        // applied to live params post-construction, which triggers a rebuild via
-        // evaluation_build_config_from_live.
-        distance_model_metric: crate::spatial_vbap::DistanceMetric::default(),
-        distance_diffuse_metric: crate::spatial_vbap::DistanceMetric::default(),
-    }
-}
-
-/// Format-agnostic spatial metadata for one audio channel (bed or object).
-///
-/// The renderer accepts this type instead of any format-specific event type
-/// (e.g. `damf::Event`), so that any spatial audio source (ADM, …)
-/// can feed the renderer without a format dependency.
-///
-/// All fields except `channel_idx` and `is_bed` are `Option` — `None` means
-/// "keep the previously cached value for this channel".
-#[derive(Debug, Clone)]
-pub struct SpatialChannelEvent {
-    /// PCM channel index in the interleaved input buffer.
-    pub channel_idx: usize,
-    /// `true` → direct speaker routing (bed); `false` → VBAP spatialization (object).
-    pub is_bed: bool,
-    /// Channel gain in dB (`None` = unchanged).
-    pub gain_db: Option<i8>,
-    /// Ramp duration in audio frames (`None` = unchanged).
-    pub ramp_length: Option<u32>,
-    /// Object spatial extent per axis (w, d, h), each in [0.0, 1.0]
-    /// (`None` = unchanged). `[0.0, 0.0, 0.0]` denotes a point source.
-    pub size: Option<[f32; 3]>,
-    /// Target position in ADM Cartesian coordinates [x, y, z] (`None` = unchanged).
-    /// x ∈ [-1, 1] left/right · y ∈ [-1, 1] back/front · z ∈ [-1, 1] floor/ceiling.
-    pub position: Option<[f64; 3]>,
-    /// Sample index within the access unit where this event takes effect.
-    pub sample_pos: Option<u64>,
-}
-
-/// Per-channel state for movement detection and gain ramping
-#[derive(Clone)]
-struct ChannelState {
-    /// Gain in dB
-    gain_db: i8,
-
-    ramp: ChannelRampState,
-
-    /// `RampMode::Interp` only: this channel's destination band gains from the
-    /// previous block, reused as the start of the next block's interpolation.
-    /// Empty until the first block for this channel.
-    interp_prev_gains: Vec<Gains>,
-}
-
-impl Default for ChannelState {
-    fn default() -> Self {
-        Self {
-            gain_db: -128, // -inf dB (muted)
-            ramp: ChannelRampState::default(),
-            interp_prev_gains: Vec::new(),
-        }
-    }
-}
-
-/// VBAP engine for one frequency band.
-///
-/// Built with the same parameters as the main renderer (`table_mode`, `allow_negative_z`,
-/// `position_interpolation`).  `compute_gains` always returns full-size `Gains`
-/// (`num_speakers` entries) with zeros for speakers outside this band, enabling
-/// uniform SIMD-friendly accumulation in the render loop.
-struct BandRenderer {
-    /// Global speaker indices for the speakers in this band.
-    speaker_indices: Vec<usize>,
-    /// Full speaker count — size of the returned `Gains`.
-    num_speakers: usize,
-    /// Band-restricted render topology (its `backend` is the prepared engine).
-    /// Stored (rather than just the engine) so a geometry-unchanged refresh can
-    /// reuse its gain model via `build_topology_reusing`. Always `Some` for a band
-    /// with ≥1 speaker — 1–2 speakers use the degenerate-VBAP `FewSpeakerBackend`,
-    /// ≥3 the full panner. `None` only for an empty (coverage-gap) band, which
-    /// renders silence.
-    topology: Option<Arc<RenderTopology>>,
-    /// `true` when this band covers every speaker in identity order
-    /// (`speaker_indices == 0..num_speakers`) — the no-crossover case. Then the
-    /// band-local gains are already full-size and the scatter is skipped.
-    is_identity: bool,
-}
-
-impl BandRenderer {
-    fn from_band(
-        band: &FreqBand,
-        layout: &crate::speaker_layout::SpeakerLayout,
-        num_speakers: usize,
-        control: &Arc<RendererControl>,
-        prev: Option<&BandRenderer>,
-    ) -> Result<Self> {
-        let speaker_indices = band.speaker_indices.clone();
-        // Log which speakers fall into this crossover band and its frequency range.
-        // A speaker is in band [lo, hi) when its usable range overlaps it, so the
-        // sub band [0, lo) holds every *full-range* spatializable speaker (those
-        // without a `freq_low` cutoff); band-limited speakers join higher bands.
-        // (LFE / non-spatializable speakers are excluded from every band.)
-        let band_high = if band.high_hz.is_finite() {
-            format!("{:.0}", band.high_hz)
-        } else {
-            "inf".to_string()
-        };
-        let band_speaker_names: Vec<&str> = speaker_indices
-            .iter()
-            .map(|&idx| layout.speakers[idx].name.as_str())
-            .collect();
-        if speaker_indices.is_empty() {
-            log::warn!(
-                "Crossover band [{:.0}, {}) Hz has no speakers — a coverage gap, content in \
-                 this range is dropped",
-                band.low_hz,
-                band_high,
-            );
-        } else {
-            log::info!(
-                "Crossover band [{:.0}, {}) Hz: {} speakers {:?}",
-                band.low_hz,
-                band_high,
-                speaker_indices.len(),
-                band_speaker_names,
-            );
-        }
-        let topology = if !speaker_indices.is_empty() {
-            let band_layout = crate::speaker_layout::SpeakerLayout {
-                radius_m: layout.radius_m,
-                speakers: speaker_indices
-                    .iter()
-                    .map(|&idx| layout.speakers[idx].clone())
-                    .collect(),
-            };
-            let plan = control
-                .prepare_topology_rebuild_for_layout(band_layout)
-                .ok_or_else(|| anyhow::anyhow!("failed to prepare band topology rebuild"))?;
-            // Reuse the previous band's geometry when it covers the same speakers
-            // and the geometry generation is unchanged: `build_topology_reusing`
-            // then re-wraps the model (no re-triangulation).
-            let prev_topology = prev
-                .filter(|p| p.speaker_indices == speaker_indices)
-                .and_then(|p| p.topology.as_deref());
-            Some(Arc::new(plan.build_topology_reusing(prev_topology)?))
-        } else {
-            None
-        };
-
-        let is_identity = speaker_indices.len() == num_speakers
-            && speaker_indices.iter().enumerate().all(|(i, &idx)| i == idx);
-
-        Ok(Self {
-            speaker_indices,
-            num_speakers,
-            topology,
-            is_identity,
-        })
-    }
-
-    /// The prepared engine for this band, when it has one (≥3 speakers).
-    fn engine(&self) -> Option<&Arc<PreparedRenderEngine>> {
-        self.topology.as_ref().map(|t| &t.backend)
-    }
-
-    /// Compute VBAP gains for this band at `position`.
-    ///
-    /// Returns full-size `Gains` (length = `num_speakers`): band speakers get their
-    /// VBAP gain, all other speakers get 0.
-    fn compute_gains(
-        &self,
-        render_params: crate::ramp_strategy::RampRenderParams,
-        position: [f64; 3],
-        event_size: [f32; 3],
-    ) -> crate::spatial_vbap::Gains {
-        let req = render_params.render_request_for_event(position, event_size);
-        let n = self.speaker_indices.len();
-        let band_gains = match self.engine() {
-            Some(engine) => engine.compute_gains(&req).gains,
-            // Only an empty (coverage-gap) band has no engine; it renders silence.
-            // Bands with 1–2 speakers carry a `FewSpeakerBackend` engine.
-            None => crate::spatial_vbap::Gains::zeroed(n),
-        };
-        // No crossover: band gains are already full-size and in speaker order, so
-        // return them directly instead of zeroing + scattering into a fresh Gains.
-        if self.is_identity {
-            return band_gains;
-        }
-        // Scatter band-local gains into a full-size vector.
-        let mut full = crate::spatial_vbap::Gains::zeroed(self.num_speakers);
-        for (gi, &g) in band_gains.iter().enumerate() {
-            full.set(self.speaker_indices[gi], g);
-        }
-        full
-    }
-}
-
-/// Split one sample into frequency bands.
-///
-/// When a crossover filter bank is active, runs the sample through the LR4 chain.
-/// Otherwise returns a 1-band passthrough so the caller's band loop is identical.
-#[inline]
-fn split_bands(
-    raw: f32,
-    filter_bank: &Option<LR4CrossoverBank>,
-    states: Option<&mut [BiquadState]>,
-) -> crate::crossover::SmallBands {
-    match (filter_bank.as_ref(), states) {
-        (Some(fb), Some(s)) => fb.process_sample(raw, s),
-        _ => crate::crossover::SmallBands::single(raw),
-    }
-}
+mod components;
+use components::{BandRenderer, ChannelState, evaluation_build_config, split_bands};
+pub use components::{RenderedFrame, SpatialChannelEvent};
 
 /// Snapshot of `LiveParams` taken at the start of each render frame.
 ///
@@ -2082,538 +1811,4 @@ impl SpatialRenderer {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    /// The unified multi-band cartesian table must render bit-equivalently to the
-    /// per-band path it replaces. Build two identical crossover renderers, force
-    /// one onto the per-band path (`unified_table = None`), feed both the same
-    /// frame, and require matching output.
-    #[test]
-    fn unified_crossover_matches_per_band() {
-        fn build() -> SpatialRenderer {
-            let mut layout = SpeakerLayout::preset("7.1.4").unwrap();
-            for (sp, cutoff) in layout.speakers.iter_mut().zip([80.0, 200.0, 500.0]) {
-                sp.freq_low = Some(cutoff);
-            }
-            SpatialRenderer::new(
-                layout,
-                48_000,
-                1,
-                1,
-                0.0,
-                2.0,
-                VbapTableMode::Cartesian {
-                    x_size: 21,
-                    y_size: 21,
-                    z_size: 9,
-                    z_neg_size: 9,
-                },
-                false,
-                true, // position interpolation → trilinear lookup + per-sample motion
-                DistanceModel::Linear,
-                false,
-                1.0,
-                1.0,
-                0.0,
-                1.0,
-                false,
-                [1.0, 2.0, 0.5],
-                2.0,
-                0.5,
-                0.0,
-                0.0,
-                false,
-                false,
-                false,
-                1.0,
-                1.0,
-                PreferredEvaluationMode::PrecomputedCartesian,
-                LiveEvaluationMode::PrecomputedCartesian,
-                21,
-                21,
-                9,
-                9,
-            )
-            .unwrap()
-        }
-
-        let mut unified = build();
-        assert!(
-            unified.unified_table.is_some(),
-            "crossover layout should build a unified table"
-        );
-        let mut per_band = build();
-        per_band.unified_table = None;
-
-        let pcm: Vec<f32> = (0..40).map(|i| (i * 7 % 13) as f32 / 13.0 - 0.5).collect();
-        let event = vec![SpatialChannelEvent {
-            channel_idx: 0,
-            is_bed: false,
-            gain_db: Some(0),
-            ramp_length: Some(40),
-            size: Some([0.0, 0.0, 0.0]),
-            position: Some([0.3, -0.2, 0.4]),
-            sample_pos: Some(0),
-        }];
-
-        let a = unified
-            .render_frame(&pcm, 1, &event, Vec::new(), false)
-            .unwrap();
-        let b = per_band
-            .render_frame(&pcm, 1, &event, Vec::new(), false)
-            .unwrap();
-        assert_eq!(a.samples.len(), b.samples.len());
-        let max_diff = a
-            .samples
-            .iter()
-            .zip(&b.samples)
-            .map(|(x, y)| (x - y).abs())
-            .fold(0.0f32, f32::max);
-        assert!(
-            max_diff < 1e-6,
-            "unified vs per-band output mismatch: max diff {max_diff}"
-        );
-    }
-
-    /// Polar counterpart of `unified_crossover_matches_per_band`: the unified
-    /// multi-band POLAR table must render bit-equivalently to the per-band polar
-    /// path. Same crossover layout, but a precomputed polar evaluator.
-    #[test]
-    fn unified_polar_matches_per_band() {
-        fn build() -> SpatialRenderer {
-            let mut layout = SpeakerLayout::preset("7.1.4").unwrap();
-            for (sp, cutoff) in layout.speakers.iter_mut().zip([80.0, 200.0, 500.0]) {
-                sp.freq_low = Some(cutoff);
-            }
-            SpatialRenderer::new(
-                layout,
-                48_000,
-                1,
-                1,
-                0.0,
-                2.0,
-                VbapTableMode::Polar,
-                false,
-                true, // position interpolation → trilinear lookup + per-sample motion
-                DistanceModel::Linear,
-                false,
-                1.0,
-                1.0,
-                0.0,
-                1.0,
-                false,
-                [1.0, 2.0, 0.5],
-                2.0,
-                0.5,
-                0.0,
-                0.0,
-                false,
-                false,
-                false,
-                1.0,
-                1.0,
-                PreferredEvaluationMode::PrecomputedPolar,
-                LiveEvaluationMode::PrecomputedPolar,
-                31,
-                31,
-                15,
-                15,
-            )
-            .unwrap()
-        }
-
-        let mut unified = build();
-        assert!(
-            unified.unified_table.is_some(),
-            "polar crossover layout should build a unified table"
-        );
-        let mut per_band = build();
-        per_band.unified_table = None;
-
-        let pcm: Vec<f32> = (0..40).map(|i| (i * 7 % 13) as f32 / 13.0 - 0.5).collect();
-        let event = vec![SpatialChannelEvent {
-            channel_idx: 0,
-            is_bed: false,
-            gain_db: Some(0),
-            ramp_length: Some(40),
-            size: Some([0.0, 0.0, 0.0]),
-            position: Some([0.3, -0.2, 0.4]),
-            sample_pos: Some(0),
-        }];
-
-        let a = unified
-            .render_frame(&pcm, 1, &event, Vec::new(), false)
-            .unwrap();
-        let b = per_band
-            .render_frame(&pcm, 1, &event, Vec::new(), false)
-            .unwrap();
-        assert_eq!(a.samples.len(), b.samples.len());
-        let max_diff = a
-            .samples
-            .iter()
-            .zip(&b.samples)
-            .map(|(x, y)| (x - y).abs())
-            .fold(0.0f32, f32::max);
-        assert!(
-            max_diff < 1e-6,
-            "unified polar vs per-band output mismatch: max diff {max_diff}"
-        );
-    }
-
-    /// A crossover band with only 1–2 speakers used to have no engine (hardcoded
-    /// equal-power), which disabled the unified table for the whole crossover.
-    /// Now such a band carries a `FewSpeakerBackend`, so the unified table builds
-    /// and must stay bit-equivalent to the per-band path. Here the top band keeps
-    /// exactly 2 spatializable speakers (pairwise-VBAP fallback).
-    #[test]
-    fn unified_table_with_two_speaker_fallback_band() {
-        fn build() -> SpatialRenderer {
-            let mut layout = SpeakerLayout::preset("7.1.4").unwrap();
-            // Cut all spatializable speakers at 200 Hz except the first two, so the
-            // [200, ∞) band has exactly 2 speakers (a fallback band) and the
-            // [0, 200) band keeps the rest (a normal ≥3 VBAP band).
-            let mut kept = 0;
-            for sp in layout.speakers.iter_mut() {
-                if !sp.spatialize {
-                    continue;
-                }
-                if kept < 2 {
-                    kept += 1;
-                    continue;
-                }
-                sp.freq_high = Some(200.0);
-            }
-            SpatialRenderer::new(
-                layout,
-                48_000,
-                1,
-                1,
-                0.0,
-                2.0,
-                VbapTableMode::Cartesian {
-                    x_size: 21,
-                    y_size: 21,
-                    z_size: 9,
-                    z_neg_size: 9,
-                },
-                false,
-                true,
-                DistanceModel::Linear,
-                false,
-                1.0,
-                1.0,
-                0.0,
-                1.0,
-                false,
-                [1.0, 2.0, 0.5],
-                2.0,
-                0.5,
-                0.0,
-                0.0,
-                false,
-                false,
-                false,
-                1.0,
-                1.0,
-                PreferredEvaluationMode::PrecomputedCartesian,
-                LiveEvaluationMode::PrecomputedCartesian,
-                21,
-                21,
-                9,
-                9,
-            )
-            .unwrap()
-        }
-
-        let mut unified = build();
-        assert!(
-            unified.unified_table.is_some(),
-            "a 2-speaker fallback band must not disable the unified table"
-        );
-        let mut per_band = build();
-        per_band.unified_table = None;
-
-        let pcm: Vec<f32> = (0..40).map(|i| (i * 7 % 13) as f32 / 13.0 - 0.5).collect();
-        let event = vec![SpatialChannelEvent {
-            channel_idx: 0,
-            is_bed: false,
-            gain_db: Some(0),
-            ramp_length: Some(40),
-            size: Some([0.0, 0.0, 0.0]),
-            position: Some([0.3, -0.2, 0.4]),
-            sample_pos: Some(0),
-        }];
-
-        let a = unified
-            .render_frame(&pcm, 1, &event, Vec::new(), false)
-            .unwrap();
-        let b = per_band
-            .render_frame(&pcm, 1, &event, Vec::new(), false)
-            .unwrap();
-        assert_eq!(a.samples.len(), b.samples.len());
-        let max_diff = a
-            .samples
-            .iter()
-            .zip(&b.samples)
-            .map(|(x, y)| (x - y).abs())
-            .fold(0.0f32, f32::max);
-        assert!(
-            max_diff < 1e-6,
-            "unified vs per-band output mismatch (fallback band): max diff {max_diff}"
-        );
-    }
-
-    /// An evaluation-mode change must reuse the triangulated gain model (the
-    /// geometry is mode-independent), rebuilding only the evaluation wrapper. A
-    /// geometry change (bumped generation) must rebuild the model. Verified via
-    /// `Arc::ptr_eq` on the decorated model.
-    #[test]
-    fn eval_mode_change_reuses_geometry() {
-        let layout = SpeakerLayout::preset("7.1.4").unwrap();
-        let r = SpatialRenderer::new(
-            layout,
-            48_000,
-            1,
-            1,
-            0.0,
-            2.0,
-            VbapTableMode::Cartesian {
-                x_size: 21,
-                y_size: 21,
-                z_size: 9,
-                z_neg_size: 9,
-            },
-            false,
-            true,
-            DistanceModel::Linear,
-            false,
-            1.0,
-            1.0,
-            0.0,
-            1.0,
-            false,
-            [1.0, 2.0, 0.5],
-            2.0,
-            0.5,
-            0.0,
-            0.0,
-            false,
-            false,
-            false,
-            1.0,
-            1.0,
-            PreferredEvaluationMode::PrecomputedCartesian,
-            LiveEvaluationMode::PrecomputedCartesian,
-            21,
-            21,
-            9,
-            9,
-        )
-        .unwrap();
-        let control = r.renderer_control();
-        let topo0 = control.active_topology();
-        let model0 = topo0
-            .backend
-            .decorated_model()
-            .expect("vbap backend exposes a decorated model");
-
-        // Evaluation-mode-only change: geometry generation unchanged → reuse model.
-        control
-            .live
-            .write()
-            .set_evaluation_mode(LiveEvaluationMode::Realtime);
-        let plan = control.prepare_topology_rebuild().expect("rebuild plan");
-        let reused = plan
-            .build_topology_reusing(Some(&topo0))
-            .expect("reuse build");
-        assert_eq!(
-            reused.backend.evaluation_mode(),
-            EffectiveEvaluationMode::Realtime
-        );
-        assert!(
-            Arc::ptr_eq(&model0, &reused.backend.decorated_model().unwrap()),
-            "evaluation-mode change must reuse the triangulated gain model"
-        );
-
-        // Geometry change bumps the generation → full rebuild (different model).
-        control.bump_geometry_generation();
-        let plan2 = control.prepare_topology_rebuild().expect("rebuild plan 2");
-        let rebuilt = plan2.build_topology_reusing(Some(&topo0)).expect("rebuild");
-        assert!(
-            !Arc::ptr_eq(&model0, &rebuilt.backend.decorated_model().unwrap()),
-            "a geometry change must rebuild the gain model"
-        );
-    }
-
-    #[test]
-    fn test_renderer_creation() {
-        let layout = SpeakerLayout::preset("7.1.4").unwrap();
-        let renderer = SpatialRenderer::new(
-            layout,
-            48000,
-            1,
-            1,
-            0.0,
-            2.0,
-            VbapTableMode::Polar,
-            false,
-            false,
-            DistanceModel::Linear,
-            false,
-            1.0,
-            1.0,
-            0.0,
-            1.0,
-            false,
-            [1.0, 2.0, 0.5],
-            2.0,
-            0.5,
-            0.0,
-            0.0,
-            false,
-            false,
-            false,
-            1.0,
-            1.0,
-            PreferredEvaluationMode::PrecomputedPolar,
-            LiveEvaluationMode::PrecomputedPolar,
-            31,
-            31,
-            15,
-            15,
-        );
-
-        assert!(renderer.is_ok());
-
-        let renderer = renderer.unwrap();
-        assert_eq!(renderer.num_speakers(), 12);
-    }
-
-    /// Guard rail: the four ramp modes must stay wired and each keep its own
-    /// behaviour. `Off` snaps to the target, `Frame` holds the block-start
-    /// position, `Sample` interpolates the position per sample, and `Interp`
-    /// interpolates the gains per sample from the previous block's end. We render
-    /// TWO blocks per mode with a position change in between (the first block
-    /// seeds `Interp`'s start gains, so its ramp only shows on the second) and
-    /// compare the second block: every output must be finite, non-silent, and
-    /// the modes must not collapse onto one another for a moving object.
-    #[test]
-    fn all_four_ramp_modes_render_distinctly() {
-        fn build() -> SpatialRenderer {
-            let layout = SpeakerLayout::preset("7.1.4").unwrap();
-            SpatialRenderer::new(
-                layout,
-                48_000,
-                1,
-                1,
-                0.0,
-                2.0,
-                VbapTableMode::Cartesian {
-                    x_size: 21,
-                    y_size: 21,
-                    z_size: 9,
-                    z_neg_size: 9,
-                },
-                false,
-                true, // position interpolation → trilinear lookup + per-sample motion
-                DistanceModel::Linear,
-                false,
-                1.0,
-                1.0,
-                0.0,
-                1.0,
-                false,
-                [1.0, 2.0, 0.5],
-                2.0,
-                0.5,
-                0.0,
-                0.0,
-                false,
-                false,
-                false,
-                1.0,
-                1.0,
-                PreferredEvaluationMode::PrecomputedCartesian,
-                LiveEvaluationMode::PrecomputedCartesian,
-                21,
-                21,
-                9,
-                9,
-            )
-            .unwrap()
-        }
-
-        let pcm = vec![0.5f32; 40];
-        let event_at = |position: [f64; 3]| {
-            vec![SpatialChannelEvent {
-                channel_idx: 0,
-                is_bed: false,
-                gain_db: Some(0),
-                ramp_length: Some(40),
-                size: Some([0.0, 0.0, 0.0]),
-                position: Some(position),
-                sample_pos: Some(0),
-            }]
-        };
-        let block_a = event_at([-0.7, 0.5, 0.2]);
-        let block_b = event_at([0.8, -0.6, 0.5]);
-
-        let render = |mode: RampMode| -> Vec<f32> {
-            let mut r = build();
-            r.control.live.write().ramp_mode = mode;
-            // First block establishes a position (and seeds Interp's start gains).
-            r.render_frame(&pcm, 1, &block_a, Vec::new(), false)
-                .unwrap();
-            // Second block moves the object — this is what we compare.
-            r.render_frame(&pcm, 1, &block_b, Vec::new(), false)
-                .unwrap()
-                .samples
-        };
-
-        let off = render(RampMode::Off);
-        let frame = render(RampMode::Frame);
-        let sample = render(RampMode::Sample);
-        let interp = render(RampMode::Interp);
-
-        let expected_len = 40 * 12;
-        for (name, out) in [
-            ("off", &off),
-            ("frame", &frame),
-            ("sample", &sample),
-            ("interp", &interp),
-        ] {
-            assert_eq!(out.len(), expected_len, "{name}: wrong output length");
-            assert!(
-                out.iter().all(|x| x.is_finite()),
-                "{name}: non-finite output"
-            );
-            let energy: f32 = out.iter().map(|x| x * x).sum();
-            assert!(energy > 0.0, "{name}: produced silence");
-        }
-
-        let max_diff = |a: &[f32], b: &[f32]| {
-            a.iter()
-                .zip(b)
-                .map(|(x, y)| (x - y).abs())
-                .fold(0.0f32, f32::max)
-        };
-
-        assert!(max_diff(&off, &frame) > 1e-3, "Off vs Frame collapsed");
-        assert!(max_diff(&off, &sample) > 1e-3, "Off vs Sample collapsed");
-        assert!(max_diff(&off, &interp) > 1e-3, "Off vs Interp collapsed");
-        assert!(
-            max_diff(&frame, &sample) > 1e-3,
-            "Frame vs Sample collapsed"
-        );
-        // Sample (position-space) and Interp (gain-space) interpolate the same
-        // endpoints differently, so they diverge mid-block too.
-        assert!(
-            max_diff(&sample, &interp) > 1e-3,
-            "Sample vs Interp collapsed"
-        );
-    }
-
-    // TODO: Add integration test with real spatial metadata
-    // For now, testing is done via real spatial audio content decoding
-}
+mod tests;
